requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
Mako==1.3.5
MarkupSafe==2.1.5
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.4
numpy==2.0.1
openai==1.41.0
pandas==2.2.2
pkgconfig==1.5.5
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
Werkzeug==3.0.3
WTForms==3.1.2
zipp==3.19.2
